==> 01_Tabula_muris_seurat.Rmd <==
---
title: "Tabula muris seurat"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r message=FALSE}
library(SingleCellExperiment)
library(ggplot2)
library(dplyr)
library(RColorBrewer)
library(viridis)
library(Seurat)
```

## Set requirements
```{r}
set.seed(12345)
```


## Load file
```{r}
sce <- readRDS("input/tm_droplet_limbMuscle_raw.rds")
sce
```

## Switch to seurat object
```{r}
#sce.seurat <- CreateSeuratObject(counts = counts(sce), project = "limb_muscle_atlas", min.cells = 3, min.features = 200, meta.data = colData(sce))
```

## Switch to Seurat
```{r}
sce.seurat <- as.Seurat(sce, counts = "counts", data = "counts", project = "tabula muris limb muscle")
```

## Run linear dim red
```{r}
sce.seurat <- sce.seurat %>%
  NormalizeData() %>%
  FindVariableFeatures(nfeatures = 2000) %>%
  ScaleData() %>%
  RunPCA()
```

## Define number of dims
```{r}
ElbowPlot(sce.seurat)
```

## Cluster
```{r}
sce.seurat <- sce.seurat %>%
  FindNeighbors(dims = 1:20) %>%
  FindClusters(resolution = 0.5) %>%
  RunUMAP(dims = 1:20)
```

## visualize clusters
```{r}
DimPlot(sce.seurat, reduction = "umap", label = TRUE)
DimPlot(sce.seurat, reduction = "umap", label = TRUE, group.by = "cell_ontology_class") + NoLegend() + ggsave("output/schaum/umap.originalAnnotation.png", dpi = 300, height = 6, width = 6)
```

```{r fig.height=10, fig.width=10}
FeaturePlot(sce.seurat, features = c("Cd79a", "Cd3g", "Lyz2", "Pecam1", "Pdgfra", "Myod1", "Plp1", "Myl9", "Thbs4")) + ggsave("output/schaum/cell_type_markers.png", dpi = 300, height = 10, width = 10)
FeaturePlot(sce.seurat, features = c("Cd14", "Fcer1g", "Il1b", "Ccl6")) + ggsave("output/schaum/myeloid_markers.png", dpi = 300, width = 6, height = 6)
FeaturePlot(sce.seurat, features = c("Wnt6", "Myl1", "Mmrn1", "Pecam1")) + ggsave("output/schaum/unannotated_markers.png", dpi = 300, height = 6, width = 6)
```

## Differential expression analysis
```{r}
sce.seurat.markers <- FindAllMarkers(sce.seurat, only.pos = TRUE, min.pct = 0.25, logfc.threshold = 0.25)
sce.seurat.markers %>% group_by(cluster) %>% top_n(n = 10, wt = avg_logFC)
```

```{r}
cell_states <- c("Mesenchymal stem", "Endothelial", "B", "Endothelial", "Tenocyte", "Satellite", "T",  "Mesenchymal stem", "Endothelial", "SMMC", "Monocyte", "Macrophage", "Endothelial", 13, 14, "Schwann", 16)
sce.seurat$cell_states <- sce.seurat$seurat_clusters
levels(sce.seurat$cell_states) <- cell_states

authors_annotation <- c("Mesenchymal stem", "Endothelial", "B", "Endothelial", 4, "Satellite", "T", "Mesenchymal stem", "Endothelial", 
                        9, "Macrophage_1", "Macrophage_2", "Endothelial", 13, 14, 15, 16)
sce.seurat$authors_annotation <- sce.seurat$seurat_clusters
levels(sce.seurat$authors_annotation) <- authors_annotation

#sce.seurat <- RenameIdents(sce.seurat, cell_states)
DimPlot(sce.seurat, reduction = "umap", label = TRUE, pt.size = 0.5, group.by = "authors_annotation") + NoLegend() + ggsave("output/schaum/umap_clusters_authors.png", dpi = 300, width = 6, height = 6)
DimPlot(sce.seurat, reduction = "umap", label = TRUE, pt.size = 0.5, group.by = "cell_states") + NoLegend() + ggsave("output/schaum/umap_clusters_known.png", dpi = 300, width = 6, height = 6)
```

## attach cluster idents to DE
```{r}
#sce.seurat$Idents <- Idents(sce.seurat)
sce.seurat$cluster <- sce.seurat$seurat_clusters
cluster_annotation <- sce.seurat@meta.data %>%
  select(cluster, authors_annotation, cell_states) %>%
  distinct()
sce.seurat.markers <- sce.seurat.markers %>%
  left_join(cluster_annotation, by = "cluster")
```

## Save RDS
```{r}
saveRDS(sce.seurat.markers, "output/schaum/tm_cluster_markers.rds")
saveRDS(sce.seurat, "input/tm_droplet_limbMuscle_seurat.rds")
```
